loader gcc warning fixes and avifile sync
git-svn-id: svn://svn.mplayerhq.hu/mplayer/trunk@13183 b3059339-0415-0410-9bf9-f77b7e298cf2

+++ loader/wine/vfw.h
@@ -19,17 +19,17 @@ typedef struct __attribute__((__packed__))
 #define _BITMAPINFOHEADER_
 typedef struct __attribute__((__packed__))
 {
-    int 	biSize;
-    int  	biWidth;
-    int  	biHeight;
+    long 	biSize;
+    long  	biWidth;
+    long  	biHeight;
     short 	biPlanes;
     short 	biBitCount;
-    int 	biCompression;
-    int 	biSizeImage;
-    int  	biXPelsPerMeter;
-    int  	biYPelsPerMeter;
-    int 	biClrUsed;
-    int 	biClrImportant;
+    long 	biCompression;
+    long 	biSizeImage;
+    long  	biXPelsPerMeter;
+    long  	biYPelsPerMeter;
+    long 	biClrUsed;
+    long 	biClrImportant;
 } BITMAPINFOHEADER, *PBITMAPINFOHEADER, *LPBITMAPINFOHEADER;
 typedef struct {
 	BITMAPINFOHEADER bmiHeader;
